Update sedi.R
sedi(a, b, c, d) now returns a list

the second list entry contains the value and the first list entry gives further details
"NO or negative skill" is returned if the sum of a and d is less than two
"SEDI equal to" is returned if both b and c differ from zero (==all errors differ from zero)
"SEDI converges to" is returned if both b and zero are equal to zero (== no errors)
"SEDI greater than" is returned if either b or c are equal to zero. in this case the first list entry is calculated from b+1 or c+1, respectively.